resolves issues in pr. Moves menu type definition. removes empty login page. changes classname prop. did not swap link import due to issues being caused by using link component
