+++ src/app/[locale]/(components)/footer.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Link from "next/link";
+import { Link } from "./link";
 
 export default function Footer() {
   return (
