+++ src/app/types.d.ts
@@ -4,11 +4,15 @@ interface EditorJSBlock {
   id: string;
   data: {
     text: string;
-  }
-  type: "paragraph"
+  };
+  type: "paragraph";
 }
 
 interface EditorJSObject {
   time: number;
   blocks: EditorJSBlock[];
 }
+
+export type MenuItem = NonNullable<
+  import("@/gql/sdk").IMenuFragment["items"]
+>[number];
