+++ src/core/server/getMenuTranslation.ts
@@ -1,8 +1,10 @@
-import { cache } from "react"
-import { IMenuItemFragment } from "@/gql/sdk"
+import { cache } from "react";
+import { IMenuItemFragment } from "@/gql/sdk";
 
-export const getMenuTranslation = cache(<T extends IMenuItemFragment>(item: T) => {
+export const getMenuTranslation = cache(
+  <T extends IMenuItemFragment>(item: T) => {
     return {
-        name: item.translation?.name ?? item.category?.name
-    }
-});
+      name: item.translation?.name ?? item.category?.name,
+    };
+  }
+);
