--- src/gql/fragments/headerMenuItem.gql
@@ -1,12 +1,0 @@
-fragment HeaderMenuItemFragment on MenuItem {
-  page {
-    slug
-  }
-  children {
-     {
-      page {
-        slug
-      }
-    }
-  }
-}
